+++ apps/api/Program.cs
@@ -21,6 +21,7 @@ using Mohandseto.Api.Application.AdminOrders;
 using Mohandseto.Api.Application.AdminQuotes;
 using Mohandseto.Api.Application.AdminContent;
 using Mohandseto.Api.Application.AdminInventory;
+using Mohandseto.Api.Application.AdminProcurement;
 using Mohandseto.Api.Infrastructure;
 using Serilog;
 
@@ -91,6 +92,7 @@ builder.Services.AddScoped<AdminOrderService>();
 builder.Services.AddScoped<AdminQuoteService>();
 builder.Services.AddScoped<AdminContentService>();
 builder.Services.AddScoped<AdminInventoryService>();
+builder.Services.AddScoped<AdminProcurementService>();
 builder.Services.AddHostedService<ContentDispatchWorker>();
 builder.Services.AddScoped<CustomizationService>();
 builder.Services.AddSingleton<ISmsSender, ConsoleSmsSender>();